requirements.txt
Django==2.1
django-appconf==1.0.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-imagekit==4.0.2
djangorestframework==3.9.0
pilkit==2.0
Pillow==5.3.0
python-decouple==3.1
pytz==2018.6
six==1.11.0
sqlparse==0.2.4
